+++ include/mimic++/expectation/Builder.hpp
@@ -13,9 +13,9 @@
 #include "mimic++/expectation/Common.hpp"
 #include "mimic++/expectation/Owner.hpp"
 #include "mimic++/expectation/Registry.hpp"
-#include "mimic++/expectation/policies/ArgRequirementPolicies.hpp"
 #include "mimic++/expectation/policies/ControlPolicies.hpp"
 #include "mimic++/expectation/policies/GeneralPolicies.hpp"
+#include "mimic++/expectation/policies/RequirementPolicies.hpp"
 #include "mimic++/macros/ScopedExpectation.hpp"
 #include "mimic++/matchers/Common.hpp"
 #include "mimic++/matchers/GeneralMatchers.hpp"
