+++ test/unit-tests/expectation/policies/CMakeLists.txt
@@ -4,10 +4,10 @@
 #          https://www.boost.org/LICENSE_1_0.txt)
 
 target_sources(${TARGET_NAME} PRIVATE
-    "ArgRequirementPolicies.cpp"
     "ArgumentList.cpp"
     "ControlPolicies.cpp"
     "FinalizerPolicies.cpp"
     "GeneralPolicies.cpp"
+    "RequirementPolicies.cpp"
     "SideEffectPolicies.cpp"
 )
